Bug Fix: Interrupt Scheduling
Scheduled interrupt 5mins after current time (just before going to sleep)
not for 5mins after last interrupt. This would be a problem if the Stalker
took longer than 5mins to complete one cycle. Because then the wake
interrupt would be scheduled in the past. Therefore would never happen.

Also added a bit more debugging and join limts to the rest of the examles.